+++ src/app/sites/ki-tools/components/demonstrator-example/example-card/example-card.component.ts
@@ -11,10 +11,6 @@ export class ExampleCardComponent {
   @Input() public cardTextData: KIToolsTypes.LinkCardData;
   constructor() { }
 
-  goToUrl(url: string){
-    console.log(url);
-  }
-
   getType(){
     return this.cardTextData.style;
   }